PR #12877 - never store JettyServerUpgradeRequest in FrameHandler
