+++ jetty-core/jetty-websocket/jetty-websocket-jetty-server/src/main/java/module-info.java
@@ -21,4 +21,7 @@ module org.eclipse.jetty.websocket.server
     requires transitive org.eclipse.jetty.websocket.api;
 
     exports org.eclipse.jetty.websocket.server;
+    exports org.eclipse.jetty.websocket.server.internal to
+        org.eclipse.jetty.ee10.websocket.jetty.server,
+        org.eclipse.jetty.ee11.websocket.jetty.server;
 }
