+++ jetty-ee10/jetty-ee10-websocket/jetty-ee10-websocket-jetty-server/src/main/java/module-info.java
@@ -14,13 +14,12 @@
 module org.eclipse.jetty.ee10.websocket.jetty.server
 {
     requires jakarta.servlet;
-    requires org.eclipse.jetty.websocket.core.server;
-    requires org.eclipse.jetty.websocket.common;
     requires org.eclipse.jetty.ee10.websocket.servlet;
-    requires org.slf4j;
+    requires org.eclipse.jetty.websocket.server;
+    requires org.eclipse.jetty.websocket.common;
+    requires org.eclipse.jetty.websocket.core.server;
 
     requires transitive org.eclipse.jetty.ee10.webapp;
-    requires transitive org.eclipse.jetty.websocket.api;
 
     // Only required if using JMX.
     requires static org.eclipse.jetty.jmx;
