+++ jetty-ee11/jetty-ee11-websocket/jetty-ee11-websocket-jetty-server/pom.xml
@@ -45,6 +45,10 @@
       <artifactId>jetty-websocket-jetty-common</artifactId>
     </dependency>
     <dependency>
+      <groupId>org.eclipse.jetty.websocket</groupId>
+      <artifactId>jetty-websocket-jetty-server</artifactId>
+    </dependency>
+    <dependency>
       <groupId>org.slf4j</groupId>
       <artifactId>slf4j-api</artifactId>
     </dependency>
